generalize CLI to support various ansatzes
