+++ src/deepqmc/io.py
@@ -6,7 +6,7 @@ import torch
 
 from .errors import TomlError
 from .molecule import Molecule
-from .wf import PauliNet
+from .wf import ANSATZES
 
 log = logging.getLogger(__name__)
 
@@ -14,8 +14,8 @@ __all__ = ()
 
 
 def validate_params(params):
-    REQUIRED = {'system'}
-    OPTIONAL = {'model_kwargs', 'train_kwargs', 'evaluate_kwargs'}
+    REQUIRED = {'system', 'ansatz'}
+    OPTIONAL = {'train_kwargs', 'evaluate_kwargs'} | {f'{a}_kwargs' for a in ANSATZES}
     params = set(params)
     missing = REQUIRED - params
     if missing:
@@ -47,6 +47,11 @@ def wf_from_file(workdir):
         mol = import_fullname(name)(**system)
     else:
         mol = Molecule.from_name(name, **system)
-    model_kwargs = params.pop('model_kwargs', {})
-    wf = PauliNet.from_hf(mol, workdir=workdir, **model_kwargs)
+    ansatz = params.pop('ansatz')
+    ansatz = ANSATZES[ansatz]
+    kwargs = params.pop(f'{ansatz.name}_kwargs', {})
+    if ansatz.uses_workdir:
+        assert 'workdir' not in kwargs
+        kwargs['workdir'] = workdir
+    wf = ansatz.entry(mol, **kwargs)
     return wf, params, state
