+++ src/deepqmc/utils.py
@@ -1,4 +1,8 @@
+import inspect
+
 import numpy as np
+import tomlkit
+from tomlkit.items import Comment, Trivia
 
 __all__ = ()
 
@@ -61,3 +65,48 @@ class _EnergyOffset:
 
 
 energy_offset = _EnergyOffset()
+
+
+def _get_subkwargs(func, name, mapping):
+    target = mapping[func, name]
+    target, override = target if isinstance(target, tuple) else (target, [])
+    sub_kwargs = collect_kwarg_defaults(target, mapping)
+    for x in override:
+        if isinstance(x, tuple):
+            key, val = x
+            sub_kwargs[key] = val
+        else:
+            del sub_kwargs[x]
+    return sub_kwargs
+
+
+def collect_kwarg_defaults(func, mapping):
+    kwargs = tomlkit.table()
+    for p in inspect.signature(func).parameters.values():
+        if p.name == 'kwargs':
+            assert p.default is p.empty
+            assert p.kind is inspect.Parameter.VAR_KEYWORD
+            sub_kwargs = _get_subkwargs(func, 'kwargs', mapping)
+            for item in sub_kwargs.value.body:
+                kwargs.add(*item)
+        elif p.name.endswith('_kwargs'):
+            if mapping.get((func, p.name)) is True:
+                kwargs[p.name] = p.default
+            else:
+                assert p.default is None
+                assert p.kind is inspect.Parameter.KEYWORD_ONLY
+                sub_kwargs = _get_subkwargs(func, p.name, mapping)
+                kwargs[p.name] = sub_kwargs
+        elif p.kind is inspect.Parameter.POSITIONAL_OR_KEYWORD:
+            assert p.default in (p.empty, p.default)
+        else:
+            assert p.kind is inspect.Parameter.KEYWORD_ONLY
+            if p.default is None:
+                kwargs.add(Comment(Trivia(comment=f'#: {p.name} = ...')))
+            else:
+                try:
+                    kwargs[p.name] = p.default
+                except ValueError:
+                    print(func, p.name, p.kind, p.default)
+                    raise
+    return kwargs
